+++ src/context/AppProvider.tsx
@@ -7,27 +7,13 @@ import React, {
   useReducer,
   Reducer,
   useEffect,
-  useRef,
 } from 'react';
-import {
-  Connectors,
-  ChainUnsupportedError,
-  ConnectionRejectedError,
-  ConnectorUnsupportedError,
-  useWallet,
-} from 'use-wallet';
-import MetamaskOnboarding from '@metamask/onboarding';
 import { useHistory } from 'react-router-dom';
 import { configureScope } from '@sentry/react';
 
-import { InjectedEthereum, Connector } from '../types';
-import { CHAIN_ID, NETWORK_NAMES } from '../web3/constants';
-import { CONNECTORS } from '../web3/connectors';
-import {
-  useAddInfoNotification,
-  useAddErrorNotification,
-} from './NotificationsProvider';
-import { LocalStorage, Storage } from '../localStorage';
+import { CHAIN_ID } from '../web3/constants';
+import { useAddErrorNotification } from './NotificationsProvider';
+import { useWalletAddress, useConnected, useWallet } from './OnboardProvider';
 
 export enum AccountItems {
   Notifications,
@@ -35,29 +21,11 @@ export enum AccountItems {
 }
 
 enum Actions {
-  ResetWallet,
-  ConnectWallet,
-  ConnectWalletError,
-  ConnectWalletSuccess,
-  SetWalletSubType,
   SelectMasset,
   SupportedChainSelected,
   SetOnline,
   SetAccountItem,
   ToggleAccount,
-}
-
-enum WalletConnectionStatus {
-  Disconnected,
-  Connecting,
-  Connected,
-}
-
-enum Reasons {
-  RejectedActivation = 'Wallet activation rejected',
-  UnsupportedChain = 'Unsupported network',
-  UnsupportedConnector = 'Unsupported connector',
-  Unknown = 'Unknown error',
 }
 
 export enum StatusWarnings {
@@ -66,15 +34,8 @@ export enum StatusWarnings {
 }
 
 interface State {
-  wallet: {
-    connector?: {
-      id: keyof Connectors;
-      subType?: string;
-    };
-    status: WalletConnectionStatus;
-    error: string | null;
-    supportedChain: boolean;
-  };
+  error: string | null;
+  supportedChain: boolean;
   accountItem: AccountItems | null;
   online: boolean;
   selectedMasset: string;
@@ -84,25 +45,12 @@ type Action =
   | { type: Actions.SelectMasset; payload: string }
   | { type: Actions.SetAccountItem; payload: AccountItems | null }
   | { type: Actions.ToggleAccount; payload: AccountItems }
-  | { type: Actions.ResetWallet }
   | { type: Actions.SupportedChainSelected; payload: boolean }
-  | {
-      type: Actions.ConnectWallet;
-      payload: { id: keyof Connectors; subType?: string };
-    }
-  | {
-      type: Actions.SetWalletSubType;
-      payload?: string;
-    }
-  | { type: Actions.ConnectWalletError; payload: string }
-  | { type: Actions.ConnectWalletSuccess }
   | { type: Actions.SetOnline; payload: boolean };
 
 interface Dispatch {
   closeAccount(): void;
-  connectWallet(connector: keyof Connectors, subType?: string): void;
   openWalletRedirect(redirect: string): void;
-  resetWallet(): void;
   selectMasset(massetName: string): void;
   toggleNotifications(): void;
   toggleWallet(): void;
@@ -124,66 +72,11 @@ const reducer: Reducer<State, Action> = (state, action) => {
     case Actions.SupportedChainSelected:
       return {
         ...state,
-        wallet: {
-          ...state.wallet,
-          supportedChain: action.payload,
-          error: null,
-        },
+        supportedChain: action.payload,
+        error: null,
       };
     case Actions.SelectMasset:
       return { ...state, selectedMasset: action.payload };
-    case Actions.ResetWallet:
-      return {
-        ...state,
-        wallet: {
-          ...state.wallet,
-          connector: undefined,
-          status: WalletConnectionStatus.Disconnected,
-          error: null,
-        },
-      };
-    case Actions.ConnectWallet:
-      return {
-        ...state,
-        wallet: {
-          ...state.wallet,
-          status: WalletConnectionStatus.Connecting,
-          connector: action.payload,
-          error: null,
-        },
-      };
-    case Actions.ConnectWalletError:
-      return {
-        ...state,
-        wallet: {
-          ...state.wallet,
-          status: WalletConnectionStatus.Disconnected,
-          error: action.payload,
-        },
-      };
-    case Actions.ConnectWalletSuccess:
-      return {
-        ...state,
-        wallet: {
-          ...state.wallet,
-          expanded: false,
-          status: WalletConnectionStatus.Connected,
-          error: null,
-        },
-      };
-    case Actions.SetWalletSubType:
-      return state.wallet.connector
-        ? {
-            ...state,
-            wallet: {
-              ...state.wallet,
-              connector: {
-                ...state.wallet.connector,
-                subType: action.payload,
-              },
-            },
-          }
-        : state;
     case Actions.SetOnline:
       return { ...state, online: action.payload };
     default:
@@ -192,11 +85,8 @@ const reducer: Reducer<State, Action> = (state, action) => {
 };
 
 const initialState: State = {
-  wallet: {
-    status: WalletConnectionStatus.Disconnected,
-    error: null,
-    supportedChain: true,
-  },
+  error: null,
+  supportedChain: true,
   accountItem: null,
   selectedMasset: 'mUSD',
   online: true,
@@ -205,33 +95,17 @@ const initialState: State = {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const context = createContext<[State, Dispatch]>([initialState, {}] as any);
 
-const identifyInjectedSubType = (
-  injected: InjectedEthereum,
-): string | undefined => {
-  if (((injected as unknown) as { wallet: string }).wallet === 'MEETONE') {
-    return 'meetOne';
-  }
-
-  if (injected.isMetaMask) return 'metamask';
-  if (injected.isBrave) return 'brave';
-  if (injected.isTrust) return 'trust';
-  if (injected.isDapper) return 'dapper';
-
-  return undefined;
-};
-
 /**
  * Provider for global App state and interactions.
  */
 export const AppProvider: FC<{}> = ({ children }) => {
-  const attemptedReconnect = useRef(false);
   const history = useHistory();
-  const { connect, reset, status, account, connector, connectors } = useWallet<
-    InjectedEthereum
-  >();
   const [state, dispatch] = useReducer(reducer, initialState);
-  const addInfoNotification = useAddInfoNotification();
   const addErrorNotification = useAddErrorNotification();
+  const address = useWalletAddress();
+  const wallet = useWallet();
+  const connected = useConnected();
+  const status = connected ? 'connected' : 'connecting';
 
   const closeAccount = useCallback<Dispatch['closeAccount']>(() => {
     dispatch({ type: Actions.SetAccountItem, payload: null });
@@ -264,73 +138,6 @@ export const AppProvider: FC<{}> = ({ children }) => {
     },
     [dispatch],
   );
-
-  const resetWallet = useCallback<Dispatch['resetWallet']>(() => {
-    reset();
-    dispatch({ type: Actions.ResetWallet });
-    LocalStorage.removeItem<'connectorId'>('connectorId');
-    LocalStorage.removeItem<'connector'>('connector');
-  }, [dispatch, reset]);
-
-  const connectWallet = useCallback<Dispatch['connectWallet']>(
-    (id, subType) => {
-      dispatch({ type: Actions.ConnectWallet, payload: { id, subType } });
-
-      if (id === 'injected' && subType === 'metamask') {
-        const onboarding = new MetamaskOnboarding();
-        if (onboarding.state === 'NOT_INSTALLED') {
-          onboarding.startOnboarding();
-          return;
-        }
-      }
-
-      connect(id)
-        .then(() => {
-          const _connector = CONNECTORS.find(
-            c => c.id === id && c.subType === subType,
-          );
-
-          addInfoNotification(
-            'Connected',
-            _connector ? `Connected with ${_connector.label}` : null,
-          );
-
-          dispatch({ type: Actions.ConnectWalletSuccess });
-
-          if (_connector) {
-            LocalStorage.set<Storage, 'connector'>('connector', {
-              id,
-              subType,
-            });
-          }
-        })
-        .catch(error => {
-          let reason: string;
-
-          if (error instanceof ConnectionRejectedError) {
-            reason = Reasons.RejectedActivation;
-          } else if (error instanceof ChainUnsupportedError) {
-            reason = `${Reasons.UnsupportedChain}; please connect to ${
-              NETWORK_NAMES[CHAIN_ID as keyof typeof NETWORK_NAMES]
-            }`;
-          } else if (error instanceof ConnectorUnsupportedError) {
-            reason = Reasons.UnsupportedConnector;
-          } else {
-            // eslint-disable-next-line no-console
-            console.error(error);
-            reason = Reasons.Unknown;
-          }
-
-          addErrorNotification(reason);
-          dispatch({
-            type: Actions.ConnectWalletError,
-            payload: reason,
-          });
-        });
-    },
-    [connect, addInfoNotification, addErrorNotification],
-  );
-
   const connectionListener = useCallback(() => {
     dispatch({
       type: Actions.SetOnline,
@@ -380,10 +187,9 @@ export const AppProvider: FC<{}> = ({ children }) => {
     let chainChangedListener: (chainId: number | string) => void;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const injected = (window as any).ethereum;
-
     // The network change listener is only valid when the injected connector is
     // used, or it's present but no connector is activated.
-    if (injected && (!connector || connector === 'injected')) {
+    if (injected && (!wallet || wallet.type === 'injected')) {
       chainChangedListener = chainId => {
         // `chainId` from MetaMask can't be trusted in this event
         if (!Number.isNaN(chainId as number)) {
@@ -396,20 +202,9 @@ export const AppProvider: FC<{}> = ({ children }) => {
       };
       chainChangedListener(parseInt(injected.chainId, 16));
 
-      const subType = identifyInjectedSubType(injected);
-      if (subType === 'dapper') {
-        addErrorNotification(Reasons.UnsupportedConnector);
-        dispatch({
-          type: Actions.ConnectWalletError,
-          payload: Reasons.UnsupportedConnector,
-        });
-      } else {
-        dispatch({ type: Actions.SetWalletSubType, payload: subType });
-
-        injected.on?.('chainChanged', chainChangedListener);
-        injected.on?.('networkChanged', chainChangedListener);
-        injected.autoRefreshOnNetworkChange = false;
-      }
+      injected.on?.('chainChanged', chainChangedListener);
+      injected.on?.('networkChanged', chainChangedListener);
+      injected.autoRefreshOnNetworkChange = false;
     }
 
     return () => {
@@ -418,35 +213,7 @@ export const AppProvider: FC<{}> = ({ children }) => {
         injected.removeListener?.('chainChanged', chainChangedListener);
       }
     };
-  }, [dispatch, connector, addErrorNotification]);
-
-  /**
-   * Automatically reconnect once on startup (if possible)
-   */
-  useEffect(() => {
-    if (
-      !attemptedReconnect.current &&
-      !['connected', 'connecting'].includes(status)
-    ) {
-      const { id, subType } = LocalStorage.get('connector') || {};
-      if (id) {
-        if (id === 'injected') {
-          // eslint-disable-next-line
-          (connectors.injected as any)
-            ?.web3ReactConnector?.({})
-            ?.isAuthorized?.()
-            .then((authorized: boolean) => {
-              if (authorized) {
-                connectWallet(id, subType);
-              }
-            });
-        } else {
-          connectWallet(id, subType);
-        }
-      }
-      attemptedReconnect.current = true;
-    }
-  }, [status, connectWallet, connectors.injected]);
+  }, [dispatch, wallet, addErrorNotification]);
 
   /**
    * Set then Sentry user scope when the account changes
@@ -454,14 +221,13 @@ export const AppProvider: FC<{}> = ({ children }) => {
   useEffect(() => {
     configureScope(scope => {
       scope.setUser({
-        id: account || 'NOT_CONNECTED',
+        id: address || 'NOT_CONNECTED',
       });
       scope.setTags({
         activated: (status === 'connected').toString(),
-        connector: JSON.stringify(connector),
       });
     });
-  }, [account, status, connector]);
+  }, [address, status]);
 
   return (
     <context.Provider
@@ -470,9 +236,7 @@ export const AppProvider: FC<{}> = ({ children }) => {
           state,
           {
             closeAccount,
-            connectWallet,
             openWalletRedirect,
-            resetWallet,
             selectMasset,
             toggleNotifications,
             toggleWallet,
@@ -481,9 +245,7 @@ export const AppProvider: FC<{}> = ({ children }) => {
         [
           state,
           closeAccount,
-          connectWallet,
           openWalletRedirect,
-          resetWallet,
           selectMasset,
           toggleNotifications,
           toggleWallet,
@@ -499,47 +261,20 @@ export const useAppContext = (): [State, Dispatch] => useContext(context);
 
 export const useAppState = (): State => useAppContext()[0];
 
-export const useWalletState = (): State['wallet'] => useAppState().wallet;
-
-export const useIsSupportedChain = (): boolean =>
-  useWalletState().supportedChain;
-
-export const useIsWalletConnected = (): boolean =>
-  useWalletState().status === WalletConnectionStatus.Connected;
-
-export const useIsWalletConnecting = (): boolean =>
-  useWalletState().status === WalletConnectionStatus.Connecting;
+export const useIsSupportedChain = (): boolean => useAppState().supportedChain;
 
 export const useAccountOpen = (): boolean => useAppState().accountItem !== null;
 
 export const useAccountItem = (): State['accountItem'] =>
   useAppState().accountItem;
 
-export const useWalletConnector = (): Connector | undefined => {
-  const {
-    connector: { id, subType } = { id: undefined, subType: undefined },
-  } = useWalletState();
-
-  return useMemo(
-    () => id && CONNECTORS.find(c => c.id === id && c.subType === subType),
-    [id, subType],
-  );
-};
-
 export const useAppDispatch = (): Dispatch => useAppContext()[1];
-
-export const useConnectWallet = (): Dispatch['connectWallet'] =>
-  useAppDispatch().connectWallet;
 
 export const useCloseAccount = (): Dispatch['closeAccount'] =>
   useAppDispatch().closeAccount;
 
 export const useAppStatusWarnings = (): StatusWarnings[] => {
-  const {
-    online,
-    wallet: { supportedChain },
-  } = useAppState();
-
+  const { online, supportedChain } = useAppState();
   return useMemo(() => {
     const warnings = [];
 
@@ -558,6 +293,3 @@ export const useOpenWalletRedirect = (): Dispatch['openWalletRedirect'] =>
 
 export const useToggleNotifications = (): Dispatch['toggleNotifications'] =>
   useAppDispatch().toggleNotifications;
-
-export const useResetWallet = (): Dispatch['resetWallet'] =>
-  useAppDispatch().resetWallet;