+++ src/context/DataProvider/BlockProvider.tsx
@@ -1,7 +1,7 @@
-import React, { createContext, FC, useContext, useRef } from 'react';
-import { useWallet } from 'use-wallet';
+import React, { createContext, FC, useContext, useRef, useEffect } from 'react';
 
 import { useIsIdle } from '../UserProvider';
+import { useProvider } from '../OnboardProvider';
 
 export type MaybeBlockNumber = number | undefined;
 
@@ -9,17 +9,20 @@ const ctx = createContext<MaybeBlockNumber>(undefined);
 
 export const BlockProvider: FC<{}> = ({ children }) => {
   const blockNumber = useRef<MaybeBlockNumber>();
-  const { getBlockNumber } = useWallet();
+  const provider = useProvider();
   const idle = useIsIdle();
 
-  if (!idle) {
-    // Only set the new block number when the user is active
-    // `getBlockNumber` apparently returns a string
-    const latest = getBlockNumber();
-    blockNumber.current = latest
-      ? parseInt((latest as unknown) as string, 10)
-      : undefined;
-  }
+  useEffect(() => {
+    if (!idle) {
+      // Only set the new block number when the user is active
+      // getBlockNumber apparently returns a string
+      provider?.getBlockNumber().then(latest => {
+        blockNumber.current = latest
+          ? parseInt((latest as unknown) as string, 10)
+          : undefined;
+      });
+    }
+  }, [idle, provider]);
 
   return <ctx.Provider value={blockNumber.current}>{children}</ctx.Provider>;
 };