+++ src/context/index.tsx
@@ -3,7 +3,6 @@ import { UserProvider } from './UserProvider';
 import { AppProvider } from './AppProvider';
 import { AllDataProviders } from './DataProvider';
 import { NotificationsProvider } from './NotificationsProvider';
-import { SignerProvider } from './SignerProvider';
 import { TransactionsProvider } from './TransactionsProvider';
 import { ThemeProvider } from './ThemeProvider';
 import { EarnDataProvider } from './earn/EarnDataProvider';
@@ -12,19 +11,17 @@ import { CurveProvider } from './earn/CurveProvider';
 export const Providers: FC<{}> = ({ children }) => (
   <NotificationsProvider>
     <UserProvider>
-      <SignerProvider>
-        <AllDataProviders>
-          <CurveProvider>
-            <EarnDataProvider>
-              <TransactionsProvider>
-                <AppProvider>
-                  <ThemeProvider>{children}</ThemeProvider>
-                </AppProvider>
-              </TransactionsProvider>
-            </EarnDataProvider>
-          </CurveProvider>
-        </AllDataProviders>
-      </SignerProvider>
+      <AllDataProviders>
+        <CurveProvider>
+          <EarnDataProvider>
+            <TransactionsProvider>
+              <AppProvider>
+                <ThemeProvider>{children}</ThemeProvider>
+              </AppProvider>
+            </TransactionsProvider>
+          </EarnDataProvider>
+        </CurveProvider>
+      </AllDataProviders>
     </UserProvider>
   </NotificationsProvider>
 );