+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Next
 
-- Add support for running the dApp on IPFS
+## Version 1.14.0
+
+- Implement Onboard.js for connecting wallets; this supports more wallet types
+  and reduces complexity of the UI
 
 ## Version 1.13.0